fix: 구조 나누기(최상단 Server Component -> 하단 Client/Server 나눌수 있게 컴포넌트 세분화)
